removing tokenizer for vocab and removing tokenizer code
